solution for __init__.py
